IoT/wellmind-vision-api/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Computer Vision - Core (versões flexíveis para Python 3.13)
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.26.0  # Permite versão mais recente compatível com Python 3.13

# Deep Learning - Emotion Detection
deepface
tensorflow>=2.15.0
keras>=3.0.0

# Face Detection
mtcnn

# Authentication
python-jose[cryptography]
passlib[bcrypt]

# Database
pymongo
motor

# Utilities
python-dotenv
requests

# Testing
pytest
httpx
pytest-cov
